refactor(subscriptions): Scope cancel service to activation
## Context

The service extracted for the gated-activation teardown was named as though it
owned subscription cancellation in general. It does not: cancelling a pending
subscription — when a downgrade is superseded, on upgrade, on customer
termination or plan destroy — marks the record canceled directly and never
comes near this code, which would raise on the payment rule lookup if it did.

Its invoice lookup also assumed the gating invoice is always open. A tax
provider failure leaves that invoice failed instead, a state the payment-gated
scenarios already cover, so cancelling such a subscription raised and reached
the merchant as a misleading not-found error. The timeout path raised the same
way and dead-lettered, since expiry jobs do not retry.

## Description

Move the service beside the expiry service in the activation rules namespace,
where the rest of the gated lifecycle lives, so its name describes the one
lifecycle it tears down.

Take the gating invoice whatever state the tax provider left it in and close it
either way. Leaving a failed invoice behind would let a merchant retry it after
cancelling and revive a payable invoice against a cancelled subscription, and
skipping the teardown would strand coupon credits, which are applied before
taxes are computed and therefore survive a tax failure. Tolerate the invoice
being absent, which happens between gating a subscription and billing it.

Drop the comments that restate the code and keep the two that do not: why the
subscription is re-checked under the lock, and why at most one live payment can
exist.
